Remove bundler gem from gemfile
Normally, most of the ruby environment equipped with bundler, and
having it as a hard code with specific version forces the user to
download that version, which is not necessary most of the cases.

This commit removes bundler dependency from the gemspec, so it
will work fine as usual without forcing anyone to download that
specific version.